lightSwitch
make sure you drag this among all the other game objects on the left side of the SampleScene. Should have the attachment of the toggleColorInversion.cs.

This is the hitbox that detects when a mouse click has been performed on it and switches between the "Sprite-Default" to the "New Material".